+++ packages/lib/src/dataModel/BaseDataModel.ts
@@ -1,9 +1,4 @@
-import {
-  instanceDataTypeSymbol,
-  ModelClass,
-  ModelPropsData,
-  propsDataTypeSymbol,
-} from "../modelShared/BaseModelShared"
+import { dataTypeSymbol, ModelClass } from "../modelShared/BaseModelShared"
 import { modelInfoByClass } from "../modelShared/modelInfo"
 import { getSnapshot } from "../snapshot/getSnapshot"
 import { toTreeNode } from "../tweaker/tweak"
@@ -22,16 +17,11 @@ const dataModelInstanceCache = new WeakMap<ModelClass<AnyDataModel>, WeakMap<any
  *
  * Never override the constructor, use `onLazyInit` instead.
  *
- * @typeparam PropsData Props data type.
- * @typeparam InstanceData Instace data type.
+ * @typeparam Data Props data type.
  */
-export abstract class BaseDataModel<
-  PropsData extends { [k: string]: any },
-  InstanceData extends { [k: string]: any } = PropsData
-> {
+export abstract class BaseDataModel<Data extends { [k: string]: any }> {
   // just to make typing work properly
-  [propsDataTypeSymbol]: PropsData;
-  [instanceDataTypeSymbol]: InstanceData
+  [dataTypeSymbol]: Data
 
   /**
    * Called after the instance is created when there's the first call to `fn(M, data)`.
@@ -43,7 +33,7 @@ export abstract class BaseDataModel<
    * Use it if one of the data properties matches one of the model properties/functions.
    * This also allows access to the backed values of transformed properties.
    */
-  readonly $!: PropsData
+  readonly $!: Data
 
   /**
    * Performs a type check over the model instance.
@@ -59,7 +49,7 @@ export abstract class BaseDataModel<
   /**
    * Creates an instance of a data model.
    */
-  constructor(data: PropsData) {
+  constructor(data: Data) {
     if (!isObject(data)) {
       throw failure("data models can only work over data objects")
     }
@@ -86,8 +76,7 @@ export abstract class BaseDataModel<
     const self = this as any
 
     // delete unnecessary props
-    delete self[propsDataTypeSymbol]
-    delete self[instanceDataTypeSymbol]
+    delete self[dataTypeSymbol]
 
     internalNewDataModel(this, tweakedData as any, {
       modelClass,
@@ -121,7 +110,7 @@ export const baseDataModelPropNames = new Set<keyof AnyDataModel>(["onLazyInit",
 /**
  * Any kind of data model instance.
  */
-export interface AnyDataModel extends BaseDataModel<any, any> {}
+export interface AnyDataModel extends BaseDataModel<any> {}
 
 /**
  * A data model class declaration, made of a base model and the model interface.
@@ -129,8 +118,3 @@ export interface AnyDataModel extends BaseDataModel<any, any> {}
 export type DataModelClassDeclaration<BaseModelClass, ModelInterface> = BaseModelClass & {
   (...args: any[]): ModelInterface
 }
-
-/**
- * Data type for a data model.
- */
-export type DataModelData<M extends AnyDataModel> = ModelPropsData<M>