+++ packages/lib/src/modelShared/prop.ts
@@ -1,5 +1,4 @@
 import type { O } from "ts-toolbelt"
-import type { PropTransform } from "../propTransform/propTransform"
 import type { LateTypeChecker, TypeChecker } from "../typeChecking/TypeChecker"
 import type { IsOptionalValue } from "../utils/types"
 
@@ -15,15 +14,11 @@ export interface ModelProp<
   TPropValue,
   TPropCreationValue,
   TIsOptional,
-  TInstanceValue = TPropValue,
-  TInstanceCreationValue = TPropCreationValue,
   TIsId extends boolean = false,
   THasSetter = never
 > {
-  $propValueType: TPropValue
-  $propCreationValueType: TPropCreationValue
-  $instanceValueType: TInstanceValue
-  $instanceCreationValueType: TInstanceCreationValue
+  $valueType: TPropValue
+  $creationValueType: TPropCreationValue
   $isOptional: TIsOptional
   $isId: TIsId
   $hasSetter: THasSetter
@@ -31,7 +26,6 @@ export interface ModelProp<
   defaultFn: (() => TPropValue) | typeof noDefaultValue
   defaultValue: TPropValue | typeof noDefaultValue
   typeChecker: TypeChecker | LateTypeChecker | undefined
-  transform: PropTransform<any, any> | undefined
   setter: boolean | "assign"
 
   withSetter(mode?: boolean | "assign"): ModelPropWithSetter<this>
@@ -40,7 +34,7 @@ export interface ModelProp<
 /**
  * Any model property.
  */
-export type AnyModelProp = ModelProp<any, any, any, any, any, any, any>
+export type AnyModelProp = ModelProp<any, any, any, any, any>
 
 /**
  * Model properties.
@@ -53,31 +47,20 @@ export type OptionalModelProps<MP extends ModelProps> = {
   [K in keyof MP]: MP[K]["$isOptional"] & K
 }[keyof MP]
 
-export type ModelPropsToPropsData<MP extends ModelProps> = {
-  [k in keyof MP]: MP[k]["$propValueType"]
+export type ModelPropsToData<MP extends ModelProps> = {
+  [k in keyof MP]: MP[k]["$valueType"]
 }
 
-export type ModelPropsToPropsCreationData<MP extends ModelProps> = O.Optional<
+export type ModelPropsToCreationData<MP extends ModelProps> = O.Optional<
   {
-    [k in keyof MP]: MP[k]["$propCreationValueType"]
-  },
-  OptionalModelProps<MP>
->
-
-export type ModelPropsToInstanceData<MP extends ModelProps> = {
-  [k in keyof MP]: MP[k]["$instanceValueType"]
-}
-
-export type ModelPropsToInstanceCreationData<MP extends ModelProps> = O.Optional<
-  {
-    [k in keyof MP]: MP[k]["$instanceCreationValueType"]
+    [k in keyof MP]: MP[k]["$creationValueType"]
   },
   OptionalModelProps<MP>
 >
 
 export type ModelPropsToSetter<MP extends ModelProps> = {
   [k in keyof MP as MP[k]["$hasSetter"] & `set${Capitalize<k & string>}`]: (
-    value: MP[k]["$instanceValueType"]
+    value: MP[k]["$valueType"]
   ) => void
 }
 
@@ -85,14 +68,7 @@ export type ModelPropsToSetter<MP extends ModelProps> = {
  * A property that will be used as model id, replacing $modelId.
  * Can only be used in models and there can be only one per model.
  */
-export const idProp = (Symbol("idProp") as any) as ModelProp<
-  string,
-  string,
-  string,
-  string,
-  string,
-  true
->
+export const idProp = (Symbol("idProp") as any) as ModelProp<string, string, string, true>
 
 /**
  * @ignore
@@ -102,23 +78,19 @@ export type OnlyPrimitives<T> = Exclude<T, object>
 /**
  * A model prop that maybe / maybe not is optional, depending on if the value can take undefined.
  */
-export type MaybeOptionalModelProp<TPropValue, TInstanceValue = TPropValue> = ModelProp<
+export type MaybeOptionalModelProp<TPropValue> = ModelProp<
   TPropValue,
   TPropValue,
-  IsOptionalValue<TPropValue, string, never>,
-  TInstanceValue,
-  TInstanceValue
+  IsOptionalValue<TPropValue, string, never>
 >
 
 /**
  * A model prop that is definitely optional.
  */
-export type OptionalModelProp<TPropValue, TInstanceValue = TPropValue> = ModelProp<
+export type OptionalModelProp<TPropValue> = ModelProp<
   TPropValue,
   TPropValue | null | undefined,
-  string,
-  TInstanceValue,
-  TInstanceValue | null | undefined
+  string
 >
 
 /**
@@ -176,7 +148,7 @@ export function prop<TValue>(defaultValue: OnlyPrimitives<TValue>): OptionalMode
 export function prop<TValue>(): MaybeOptionalModelProp<TValue>
 
 // base
-export function prop<TValue>(def?: any): ModelProp<TValue, any, any, TValue, any, any, any> {
+export function prop<TValue>(def?: any): ModelProp<TValue, any, any, any, any> {
   let hasDefaultValue = false
 
   // default
@@ -187,18 +159,15 @@ export function prop<TValue>(def?: any): ModelProp<TValue, any, any, TValue, any
   const isDefFn = typeof def === "function"
 
   const obj: ReturnType<typeof prop> = {
-    $propValueType: null as any,
-    $propCreationValueType: null as any,
+    $valueType: null as any,
+    $creationValueType: null as any,
     $isOptional: null as any,
-    $instanceValueType: null as any,
-    $instanceCreationValueType: null as any,
     $isId: null as never,
     $hasSetter: null as never,
 
     defaultFn: hasDefaultValue && isDefFn ? def : noDefaultValue,
     defaultValue: hasDefaultValue && !isDefFn ? def : noDefaultValue,
     typeChecker: undefined,
-    transform: undefined,
     setter: false,
 
     withSetter(mode) {