+++ packages/lib/src/propTransform/index.ts
@@ -1,6 +1,0 @@
-export { PropTransform } from "./propTransform"
-export * from "./transformArrayAsMap"
-export * from "./transformArrayAsSet"
-export * from "./transformObjectAsMap"
-export * from "./transformStringAsDate"
-export * from "./transformTimestampAsDate"