+++ packages/lib/src/tweaker/tweak.ts
@@ -72,16 +72,12 @@ function internalTweak<T>(value: T, parentPath: ParentPath<any> | undefined): T 
 
   // unsupported
   if (isMap(value)) {
-    throw failure(
-      "maps are not directly supported. consider using 'prop_mapObject' / 'tProp_mapObject' to map a '{[k: string]: V}' property, or 'prop_mapArray' / 'tProp_mapArray' to map a '[string, V][]' property instead."
-    )
+    throw failure("maps are not directly supported. consider using 'ObjectMap' / 'asMap' instead.")
   }
 
   // unsupported
   if (isSet(value)) {
-    throw failure(
-      "sets are not directly supported. consider using 'prop_setArray' / 'tProp_setArray' to map a 'V[]' property instead."
-    )
+    throw failure("sets are not directly supported. consider using 'ArraySet' / 'asSet' instead.")
   }
 
   throw failure(